f4pga/flows: split vpr to a file; create subdir 'tools'; mv nextpnr and vpr
