+++ f4pga/flows/modules/analysis.py
@@ -19,7 +19,7 @@
 
 from pathlib import Path
 
-from f4pga.flows.common import vpr_specific_values, vpr as common_vpr, VprArgs
+from f4pga.flows.tools.vpr import vpr_specific_values, vpr, VprArgs
 from f4pga.flows.module import Module, ModuleContext
 
 
@@ -42,7 +42,7 @@ class analysisModule(Module):
         build_dir = Path(ctx.takes.eblif).parent
 
         yield "Analysis with VPR..."
-        common_vpr(
+        vpr(
             "analysis",
             VprArgs(
                 share=ctx.share,
